second commit

+++ services/classe/src/main/java/com/isi/microservices/repository/ClasseRepository.java
@@ -3,6 +3,6 @@ package com.isi.microservices.repository;
 import com.isi.microservices.model.Classe;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClasseRepository extends JpaRepository<Classe,Long> {
+public interface ClasseRepository extends JpaRepository<Classe,Integer> {
 
 }
